refactor: capsule -> timeCapsule 패키지 이름 변경
